+++ src/components/settings/EngineFieldsEditor.tsx
@@ -30,6 +30,7 @@ import { FlexRow } from "ui/spacing/Spacing";
 import { useEngineFieldsDefaultValues } from "./useEngineFieldsDefaultValues";
 import ToggleButtons from "ui/form/ToggleButtons";
 import { pxToSubpx, pxToSubpxVelPrecise } from "shared/lib/helpers/subpixels";
+import AnimationStateSelect from "components/forms/AnimationStateSelect";
 
 const { editEngineFieldValue, removeEngineFieldValue } = entitiesActions;
 
@@ -305,6 +306,16 @@ export const EngineFieldInput: FC<EngineFieldInputProps> = ({
         options={options}
         allowMultiple
         allowNone
+      />
+    );
+  }
+  if (field.type === "animationstate") {
+    return (
+      <AnimationStateSelect
+        name={field.key}
+        value={String(value || "")}
+        onChange={onChange}
+        allowDefault
       />
     );
   }